requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.0
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
packaging==21.3
phonenumbers==8.12.49
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
whitenoise==6.2.0
zope.interface==5.4.0

pytest~=7.1.2